move card components to card.jsx
