+++ client/src/components/hand.jsx
@@ -1,17 +1,7 @@
 import React, { Component } from "react";
 import '../css/card.css';
 import '../css/stylesheet.css';
-
-const getColor = (input) => 'SC'.includes(input) ? 'black' : 'red';
-
-const getSuitSymbol = (input) => {
-    const suits = '♠︎ ♥︎ ♣︎ ♦︎'.split(' ');
-    const inputs = 'S H C D'.split(' ');
-    for (let i = 0; i < 4; i++){
-        if (inputs[i] === input) return suits[i];
-    }
-    return 'yikes';
-};
+import Card from "./card";
 
 class HandRankMessageContainer extends Component {
     // props used: this.props.player.handRankMessage
@@ -52,60 +42,6 @@ class PlayerNameContainer extends Component {
                 <DealerChip isDealer={this.props.player.isDealer}/>
                 <span className="username">{this.props.player.playerName}</span>: <span className="stack">{this.props.player.chips}</span><Earnings earnings={this.props.player.earnings}/>
             </div>
-        );
-    }
-}
-
-class CardBack extends Component {
-    // renders grey card back (.waiting) for players who have folded
-    // other
-    // props used: this.props.folded
-    render() {
-        return (
-            <div className={this.props.folded ? "back-card waiting": "back-card"}>
-                <div className="back-card-ring" />
-            </div>
-        );
-    }
-}
-
-class VisibleCard extends Component {
-    // props used: this.props.card
-    render() {
-        let cardRank = (this.props.card.charAt(0) === 'T') ? '10' : this.props.card.charAt(0);
-        let cardSuit = getSuitSymbol(this.props.card.charAt(1));
-        return (
-            <div>
-                <div className="card-topleft">
-                    <div className="card-corner-rank">{cardRank}</div>
-                </div>
-                <div className="card-bottomright">
-                    <div className="card-corner-suit">{cardSuit}︎</div>
-                </div>
-            </div>
-        );
-    }
-}
-
-export class Card extends Component {
-    // renders the back of the card if we do not know what it is (i.e. if this.props.card is null).
-    // renders the front of the cards if we do know what it is.
-    // greys out the card if the player has folded
-    //
-    // props used: this.props.folded, this.props.card
-    render () {
-        let cardClassName = "card";
-        if (this.props.folded) cardClassName += " folded";
-        let card;
-        if (this.props.card === null) { // if we don't know this person's cards
-            cardClassName += " black"; // idk if we need to do this if we can't see the card
-            card = <CardBack folded={this.props.folded}/>
-        } else {
-            cardClassName += " " + getColor(this.props.card.charAt(1));
-            card = <VisibleCard card={this.props.card}/>
-        }
-        return (
-            <div className={cardClassName}>{card}</div>
         );
     }
 }
